test VM partition.sh - update 5

+++ 10-before_reboot/10-partitions.sh
@@ -1,17 +1,25 @@
 #!/bin/sh
 
-echo "We are going to create a partition table. Then, we will format partitions"
 echo "We will create 3 partitions ; for booting, for swap and for the rest"
-echo "1 argument = boot partition"
-echo "2 argument = swap partition"
-echo "3 argument = rest partition"
+echo "For that, run the this script with 4 arguments :"
+echo "1) write sda or sdb"
+echo "2) write the size of boot partition. Exemple: +500M"
+echo "3) write the size of swap partition. Exemple: +17G"
+echo "4) write the size of rest partition. Exemple: +100G"
+echo
+lsblk
 
-hard_disk="/dev/sda"
-#hard_disk="/dev/sdb"
+if [ $# -ne 4 ]
+then
+	echo "Wrong arguments number. Try again !"
+	exit 1
+fi
 
-$1=partition_boot_size
-$2=partition_swap_size
-$3=partition_rest_size
+hard_disk="/dev/$1"
+hard_disk1=$hard_disk"1"
+hard_disk2=$hard_disk"2"
+hard_disk3=$hard_disk"3"
+
 
 mkfs.ext4 $hard_disk
 echo "####################################################################"
@@ -22,43 +30,52 @@ fdisk $hard_disk << EOF
 g
 w
 EOF
+echo "####################################################################"
+echo "partition table created !                                          #"
+echo "####################################################################"
 
 fdisk $hard_disk << EOF
 n
 1
 
-$partition_boot_size
+$2
 t
 1
 w
 EOF
-
+echo "####################################################################"
+echo $hard_disk1 "created !                                              #"
+echo "####################################################################"
 
 fdisk $hard_disk << EOF
 n
 2
 
-$partition_swap_size
+$3
 t
 19
 w
 EOF
+echo "####################################################################"
+echo $hard_disk2 "created !                                              #"
+echo "####################################################################"
+
 
 fdisk $hard_disk << EOF
 n
 3
 
-$partition_rest_size
+$4
 t
 20
 w
 EOF
+echo "####################################################################"
+echo $hard_disk3 "created !                                              #"
+echo "####################################################################"
+
 
 #########################################################################################
-
-#hard_disk1=$hard_disk"1"
-#hard_disk2=$hard_disk"2"
-#hard_disk3=$hard_disk"3"
 
 #mkfs.fat -F32 $hard_disk1
 #mkswap $hard_disk2
